+++ exports/light-editor.styles.js
@@ -75,10 +75,14 @@ export const styles = css`
     /* this creates line-wrapping. */
 	  word-break: break-word;
     white-space: pre-wrap;
+  }
 
+
+  :host([wrap="none"]) :is([part~="pre"], [part~="textarea"], [part~="code"], [part~="gutter"]) {
+  [wrap="none"] {
     /* This would remove line-wrapping */
-	  /* word-break: break-all; */
-    /* white-space: pre; */
+	  word-break: break-all;
+    white-space: pre;
   }
 
   /** Hide scrollbars for the gutter */